Improvements in migrations and seeders.

+++ database/migrations/2025_11_15_211635_create_accounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('accounts', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name', 45);
-            $table->integer('type_id')->index('fk_accounts_account_types1_idx');
+            $table->unsignedBigInteger('type_id')->index('fk_accounts_account_types1_idx');
             $table->unsignedInteger('currency_id')->index('fk_accounts_account_currencys1_idx');
             $table->float('initial_balance', 15);
             $table->float('current_balance', 15);
